\name{ooc-package}
\alias{ooc-package}
\docType{package}
\title{Ordered Optimal Classification}
\description{Performs Ordered Optimal Classification, an extension of Poole's (2000) nonparametric unfolding procedure for the analysis of ordinal choice data (e.g., ``Strongly Agree,'' ``Somewhat Agree,'' ``Somewhat Disagree,'' ``Strongly Disagree'') as detailed in Hare, Liu, and Lupton (2018).}
\details{
\tabular{ll}{
Package: \tab ooc\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2021-08-19\cr
License: \tab GPL-2 \cr
}}
\author{Christopher Hare, Tzu-Ping Liu, and Keith T. Poole}
\references{Poole, Keith T. 2000. ``Nonparametric Unfolding of Binary Choice Data.'' \emph{Political Analysis} 8(3): 211--237.

Hare, Christopher, Tzu-Ping Liu, and Robert N. Lupton. 2018. ``What Ordered Optimal Classification Reveals about Ideological Structure, Cleavages, and Polarization in the American Mass Public.'' \emph{Public Choice} 176: 57--78.}
\keyword{ooc}
\seealso{oc}


